Remove claude naming
