+++ site/cell-boundaries.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useMemo } from "react"
-import { computeBoundsFromCellContents } from "./computeBoundsFromCellContents"
+import { calculateCellBoundaries } from "../lib/calculateCellBoundaries"
 
 // Type definitions
 interface CellContent {
@@ -231,584 +231,6 @@ const findClosestPointOnSegmentToAnyCells = (
   })
 
   return globalClosest
-}
-
-// Helper function to convert a Point to a string key
-// Precision should be related to POINT_COMPARISON_TOLERANCE
-const pointToKey = (p: Point): string => `${p.x.toFixed(4)},${p.y.toFixed(4)}`
-
-// Helper function to create a canonical key for a segment
-const getSegmentKey = (p1: Point, p2: Point): string => {
-  // Use toFixed to handle potential floating point inaccuracies in keys
-  const p1x = parseFloat(p1.x.toFixed(4))
-  const p1y = parseFloat(p1.y.toFixed(4))
-  const p2x = parseFloat(p2.x.toFixed(4))
-  const p2y = parseFloat(p2.y.toFixed(4))
-
-  // Sort points by x, then y to ensure (A,B) and (B,A) have the same key
-  if (p1x < p2x || (p1x === p2x && p1y < p2y)) {
-    return `${p1x},${p1y}_${p2x},${p2y}`
-  } else {
-    return `${p2x},${p2y}_${p1x},${p1y}`
-  }
-}
-
-const getAngle = (p1: Point, p2: Point): number => {
-  return Math.atan2(p2.y - p1.y, p2.x - p1.x)
-}
-
-//
-// ── Rectangle helpers (edge-to-edge) ────────────────────────────
-const edgeToEdgeDistance = (a: CellContent, b: CellContent) => {
-  const dx = Math.max(a.x - (b.x + b.width), b.x - (a.x + a.width), 0)
-  const dy = Math.max(a.y - (b.y + b.height), b.y - (a.y + a.height), 0)
-  return dx + dy // Manhattan distance
-}
-
-const areAdjacent = (a: CellContent, b: CellContent, tol = 0.5) => {
-  const shareVertical =
-    (Math.abs(a.x + a.width - b.x) < tol ||
-      Math.abs(b.x + b.width - a.x) < tol) &&
-    !(a.y + a.height <= b.y || b.y + b.height <= a.y)
-
-  const shareHorizontal =
-    (Math.abs(a.y + a.height - b.y) < tol ||
-      Math.abs(b.y + b.height - a.y) < tol) &&
-    !(a.x + a.width <= b.x || b.x + b.width <= a.x)
-
-  return shareVertical || shareHorizontal
-}
-
-// Simple union-find for grouping rects
-class DSU {
-  parent: number[]
-  constructor(n: number) {
-    this.parent = Array.from({ length: n }, (_, i) => i)
-  }
-  find(i: number): number {
-    if (this.parent[i] !== i) this.parent[i] = this.find(this.parent[i])
-    return this.parent[i]
-  }
-  union(i: number, j: number) {
-    const pi = this.find(i)
-    const pj = this.find(j)
-    if (pi !== pj) this.parent[pj] = pi
-  }
-}
-
-export const calculateCellBoundaries = (
-  inputCellContents: Omit<CellContent, "cellId">[],
-  containerWidth?: number,
-  containerHeight?: number,
-): {
-  midlines: Midline[]
-  allSegments: Array<Line>
-  validSegments: Array<Line>
-  mergedRectGroups: CellContent[][]
-  cellRects: CellContent[]
-  gridRects: CellContent[]
-  outlineLines: Line[]
-  // polygons: Array<PolygonType>; // Removed
-  // mergedPolygons: Array<PolygonType>; // Removed
-} => {
-  const originalCellContents = inputCellContents.map((cellContent, index) => ({
-    ...cellContent,
-    cellId: `cell-${index}`,
-  }))
-
-  const containerBounds = computeBoundsFromCellContents(
-    originalCellContents.map((c) => ({
-      minX: c.x,
-      minY: c.y,
-      maxX: c.x + c.width,
-      maxY: c.y + c.height,
-    })),
-  )
-
-  const offsetX = containerBounds.minX
-  const offsetY = containerBounds.minY
-
-  containerWidth ??= containerBounds.maxX - containerBounds.minX
-  containerHeight ??= containerBounds.maxY - containerBounds.minY
-
-  const cellContents = originalCellContents.map((c) => ({
-    ...c,
-    x: c.x - offsetX,
-    y: c.y - offsetY,
-  }))
-
-  // Step 1: Calculate midlines between all cells
-  const midlines: Midline[] = []
-  let midlineId = 0
-
-  for (let i = 0; i < cellContents.length; i++) {
-    for (let j = i + 1; j < cellContents.length; j++) {
-      const cell1 = cellContents[i]
-      const cell2 = cellContents[j]
-
-      const cell1Right = cell1.x + cell1.width
-      const cell2Right = cell2.x + cell2.width
-      const cell1Bottom = cell1.y + cell1.height
-      const cell2Bottom = cell2.y + cell2.height
-
-      // Vertical midline (horizontal gap)
-      const hasHorizontalGap = cell1Right < cell2.x || cell2Right < cell1.x
-      if (hasHorizontalGap) {
-        let midX
-        if (cell1Right < cell2.x) {
-          midX = (cell1Right + cell2.x) / 2
-        } else {
-          midX = (cell2Right + cell1.x) / 2
-        }
-
-        midlines.push({
-          id: `midline-${midlineId++}`,
-          start: { x: midX, y: 0 },
-          end: { x: midX, y: containerHeight },
-          cellIds: [cell1.cellId, cell2.cellId],
-          type: "vertical",
-        })
-      }
-
-      // Horizontal midline (vertical gap)
-      const hasVerticalGap = cell1Bottom < cell2.y || cell2Bottom < cell1.y
-      if (hasVerticalGap) {
-        let midY
-        if (cell1Bottom < cell2.y) {
-          midY = (cell1Bottom + cell2.y) / 2
-        } else {
-          midY = (cell2Bottom + cell1.y) / 2
-        }
-
-        midlines.push({
-          id: `midline-${midlineId++}`,
-          start: { x: 0, y: midY },
-          end: { x: containerWidth, y: midY },
-          cellIds: [cell1.cellId, cell2.cellId],
-          type: "horizontal",
-        })
-      }
-    }
-  }
-
-  // Step 2: Calculate intersections between midlines
-  const intersections: Intersection[] = []
-  for (let i = 0; i < midlines.length; i++) {
-    for (let j = i + 1; j < midlines.length; j++) {
-      const intersection = lineIntersection(midlines[i], midlines[j])
-      if (intersection) {
-        intersections.push({
-          point: intersection,
-          midlineIds: [midlines[i].id, midlines[j].id],
-        })
-      }
-    }
-  }
-
-  // Step 3: Calculate segments between intersections
-  const allSegments: Line[] = []
-  let segmentId = 0
-
-  midlines.forEach((midline) => {
-    // Find all intersections on this midline
-    const midlineIntersections = intersections
-      .filter((int) => int.midlineIds.includes(midline.id))
-      .map((int) => int.point)
-      .sort((a, b) => {
-        if (midline.type === "vertical") {
-          return a.y - b.y
-        } else {
-          return a.x - b.x
-        }
-      })
-
-    // Add start and end points
-    const allPoints = [midline.start, ...midlineIntersections, midline.end]
-
-    // Create segments between consecutive points
-    for (let i = 0; i < allPoints.length - 1; i++) {
-      const segment: Line = {
-        id: `segment-${segmentId++}`,
-        start: allPoints[i],
-        end: allPoints[i + 1],
-        fromCellIds: midline.cellIds,
-        distanceToAnyCell: segmentDistanceToAnyCell(
-          allPoints[i],
-          allPoints[i + 1],
-          cellContents,
-        ),
-      }
-      allSegments.push(segment)
-    }
-  })
-
-  // Step 4: Remove segments that intersect cell content (invalid segments)
-  const validSegments = allSegments.filter((segment) => {
-    // Check if the entire segment intersects any cell
-    return !cellContents.some((cell) => {
-      return lineIntersectsRectangle(segment.start, segment.end, cell)
-    })
-  })
-
-  // ── NEW: build grid rectangles from VALID segments ───────────
-  // 1. Collect unique x positions of vertical segments + container edges
-  const verticalXs = new Set<number>([0, containerWidth])
-  // 2. Collect unique y positions of horizontal segments + container edges
-  const horizontalYs = new Set<number>([0, containerHeight])
-
-  validSegments.forEach((seg) => {
-    const isVertical = Math.abs(seg.start.x - seg.end.x) < 0.001
-    if (isVertical) {
-      verticalXs.add(seg.start.x)
-    } else {
-      horizontalYs.add(seg.start.y)
-    }
-  })
-
-  const xs = Array.from(verticalXs).sort((a, b) => a - b)
-  const ys = Array.from(horizontalYs).sort((a, b) => a - b)
-
-  // --- Rect that tightly follows the grid around each cell ------------
-  const cellContainingRectMap = new Map<string, CellContent>()
-  const cellContainingRects: CellContent[] = []
-
-  cellContents.forEach((cell, idx) => {
-    // left / right grid lines that bound the cell
-    // xs and ys are sorted and include 0 and containerWidth/Height respectively.
-    const minXGrid = xs[0]
-    const maxXGrid = xs[xs.length - 1]
-    const minYGrid = ys[0]
-    const maxYGrid = ys[ys.length - 1]
-
-    let left = xs.filter((v) => v <= cell.x).pop()
-    if (left === undefined) left = minXGrid
-
-    let right = xs.find((v) => v >= cell.x + cell.width)
-    if (right === undefined) right = maxXGrid
-
-    let top = ys.filter((v) => v <= cell.y).pop()
-    if (top === undefined) top = minYGrid
-
-    let bot = ys.find((v) => v >= cell.y + cell.height)
-    if (bot === undefined) bot = maxYGrid
-
-    const rect: CellContent = {
-      cellId: `contain-${cell.cellId}`,
-      x: left,
-      y: top,
-      width: Math.max(0, right - left),
-      height: Math.max(0, bot - top),
-    }
-    cellContainingRectMap.set(cell.cellId, rect)
-    cellContainingRects.push(rect)
-  })
-
-  // helper: axis-aligned rectangle overlap
-  const rectsOverlap = (a: CellContent, b: CellContent) =>
-    a.x < b.x + b.width &&
-    a.x + a.width > b.x &&
-    a.y < b.y + b.height &&
-    a.y + a.height > b.y
-
-  // 3. Build every cell in the grid; drop any that intersect an input cell
-  const gridRects: CellContent[] = []
-  let gridRectId = 0
-
-  for (let xi = 0; xi < xs.length - 1; xi++) {
-    const x0 = xs[xi]
-    const x1 = xs[xi + 1]
-    if (x1 - x0 <= 0) continue
-
-    for (let yi = 0; yi < ys.length - 1; yi++) {
-      const y0 = ys[yi]
-      const y1 = ys[yi + 1]
-      if (y1 - y0 <= 0) continue
-
-      const candidate: CellContent = {
-        cellId: `gridRect-${gridRectId++}`,
-        x: x0,
-        y: y0,
-        width: x1 - x0,
-        height: y1 - y0,
-      }
-
-      // skip if candidate overlaps any original cell
-      if (cellContents.some((c) => rectsOverlap(candidate, c))) continue
-
-      gridRects.push(candidate)
-    }
-  }
-
-  // add the containing rects to the grid set
-  cellContainingRects.forEach((r) => {
-    const key = `${r.x},${r.y},${r.width},${r.height}`
-    if (
-      !gridRects.some((g) => `${g.x},${g.y},${g.width},${g.height}` === key)
-    ) {
-      gridRects.push(r)
-    }
-  })
-
-  // ── Step 5: merge GRID rects outward from each cell ───────────────
-  // One cell ⇒ one merged-rect group.  Start with the rect that
-  // contains the cell. Then, iteratively merge adjacent grid rects based on the
-  // minimum distanceToAnyCell of their bounding segments.
-
-  interface WorkRect extends CellContent {
-    merged: boolean
-    groupId: number | null
-    minBoundingSegmentDistance: number
-  }
-
-  // Working copy of every grid rect with merge flags and distance metric
-  const workRects: WorkRect[] = gridRects.map((r) => {
-    let minDistanceForRect = Infinity
-    const { x, y, width, height } = r
-
-    validSegments.forEach((segment) => {
-      const sx1 = Math.min(segment.start.x, segment.end.x)
-      const sx2 = Math.max(segment.start.x, segment.end.x)
-      const sy1 = Math.min(segment.start.y, segment.end.y)
-      const sy2 = Math.max(segment.start.y, segment.end.y)
-
-      const segmentIsHorizontal =
-        Math.abs(segment.start.y - segment.end.y) < POINT_COMPARISON_TOLERANCE
-      const segmentIsVertical =
-        Math.abs(segment.start.x - segment.end.x) < POINT_COMPARISON_TOLERANCE
-
-      let onBoundary = false
-      if (segmentIsHorizontal) {
-        if (
-          (Math.abs(sy1 - y) < POINT_COMPARISON_TOLERANCE ||
-            Math.abs(sy1 - (y + height)) < POINT_COMPARISON_TOLERANCE) &&
-          sx1 < x + width - POINT_COMPARISON_TOLERANCE &&
-          sx2 > x + POINT_COMPARISON_TOLERANCE && // ensure segment x-range overlaps rect x-range with positive length
-          Math.max(sx1, x) + POINT_COMPARISON_TOLERANCE <
-            Math.min(sx2, x + width)
-        ) {
-          onBoundary = true
-        }
-      } else if (segmentIsVertical) {
-        if (
-          (Math.abs(sx1 - x) < POINT_COMPARISON_TOLERANCE ||
-            Math.abs(sx1 - (x + width)) < POINT_COMPARISON_TOLERANCE) &&
-          sy1 < y + height - POINT_COMPARISON_TOLERANCE &&
-          sy2 > y + POINT_COMPARISON_TOLERANCE && // ensure segment y-range overlaps rect y-range with positive length
-          Math.max(sy1, y) + POINT_COMPARISON_TOLERANCE <
-            Math.min(sy2, y + height)
-        ) {
-          onBoundary = true
-        }
-      }
-
-      if (onBoundary && segment.distanceToAnyCell !== undefined) {
-        minDistanceForRect = Math.min(
-          minDistanceForRect,
-          segment.distanceToAnyCell,
-        )
-      }
-    })
-
-    return {
-      ...r,
-      merged: false,
-      groupId: null,
-      minBoundingSegmentDistance: minDistanceForRect,
-    }
-  })
-
-  // 5-a Mark the “key rects” (the cell-containing rects) as the seeds
-  cellContainingRects.forEach((contRect, idx) => {
-    const wr = workRects.find(
-      (w) =>
-        pointsEqual({ x: w.x, y: w.y }, { x: contRect.x, y: contRect.y }) &&
-        Math.abs(w.width - contRect.width) < POINT_COMPARISON_TOLERANCE &&
-        Math.abs(w.height - contRect.height) < POINT_COMPARISON_TOLERANCE,
-    )
-    if (wr) {
-      wr.merged = true
-      wr.groupId = idx // one group per cell
-    }
-  })
-
-  // 5-b Iterative grow-out: assign each rect to an adjacent merged group,
-  // prioritizing rects with lower minBoundingSegmentDistance.
-  // Skipped rects are re-evaluated in subsequent passes.
-  let unmergedRectsProcessingList: WorkRect[] = workRects
-    .filter((r) => !r.merged)
-    .sort((a, b) => a.minBoundingSegmentDistance - b.minBoundingSegmentDistance)
-
-  let keepProcessing = true
-  while (keepProcessing && unmergedRectsProcessingList.length > 0) {
-    keepProcessing = false // Assume no merges in this pass initially
-    const remainingAfterPass: WorkRect[] = []
-
-    for (const rectToConsider of unmergedRectsProcessingList) {
-      // Find the current state of this rect from the main workRects array
-      const currentRectState = workRects.find(
-        (wr) => wr.cellId === rectToConsider.cellId,
-      )
-
-      // If rect somehow already processed or not found, skip
-      if (!currentRectState || currentRectState.merged) {
-        if (currentRectState && currentRectState.merged && !keepProcessing) {
-          // If it got merged by another rect earlier in this same pass, ensure keepProcessing is true
-          // This can happen if rect A merges, then rect B (neighbour of A) is processed.
-          // However, the main check for keepProcessing is if a rect *changes* its state to merged.
-        }
-        continue
-      }
-
-      const neighbours = workRects.filter(
-        (other) => other.merged && areAdjacent(currentRectState, other),
-      )
-
-      if (neighbours.length > 0) {
-        currentRectState.merged = true
-        keepProcessing = true // A merge occurred, so at least one more pass might be needed
-
-        if (neighbours.length === 1) {
-          currentRectState.groupId = neighbours[0].groupId
-        } else {
-          // Tie-breaking: find best group to join based on distance to original cell
-          let bestGroupId = null
-          let minDistanceToOriginalCell = Infinity
-          for (const neighbour of neighbours) {
-            if (neighbour.groupId === null) continue
-
-            const originalCellForGroup = cellContents[neighbour.groupId]
-            const distance = edgeToEdgeDistance(
-              currentRectState,
-              originalCellForGroup,
-            )
-
-            if (distance < minDistanceToOriginalCell) {
-              minDistanceToOriginalCell = distance
-              bestGroupId = neighbour.groupId
-            } else if (distance === minDistanceToOriginalCell) {
-              // Further tie-breaking: prefer smaller group ID (arbitrary but consistent)
-              if (
-                bestGroupId === null ||
-                (neighbour.groupId !== null && neighbour.groupId < bestGroupId)
-              ) {
-                bestGroupId = neighbour.groupId
-              }
-            }
-          }
-          currentRectState.groupId = bestGroupId
-        }
-      } else {
-        // Cannot merge this rect in this pass, add to list for next pass
-        remainingAfterPass.push(rectToConsider)
-      }
-    }
-    // Prepare for next pass with only rects that couldn't be merged, re-sorted
-    unmergedRectsProcessingList = remainingAfterPass.sort(
-      (a, b) => a.minBoundingSegmentDistance - b.minBoundingSegmentDistance,
-    )
-  }
-
-  // 5-d  collect final groups – first element MUST be the cell-containing rect,
-  //      followed by the original cell rect and the rest of the merged grid rects
-  const mergedRectGroups: CellContent[][] = []
-  cellContainingRects.forEach((contRect, idx) => {
-    const groupRects = workRects
-      .filter((r) => r.merged && r.groupId === idx)
-      .map(({ merged, groupId, ...plain }) => plain) // strip helper props
-    const cellRect = cellContents[idx] // original cell rect
-    mergedRectGroups.push([
-      contRect,
-      cellRect,
-      ...groupRects.filter((r) => r.cellId !== contRect.cellId),
-    ])
-  })
-
-  // ── Step 6:  outlines between DIFFERENT merged-rect groups ──────────────
-  const outlineLines: Line[] = []
-  let outlineLineId = 0
-  const TOL = POINT_COMPARISON_TOLERANCE
-
-  // Only consider grid rects that were merged into some group
-  const groupedRects = workRects.filter(
-    (r): r is WorkRect & { groupId: number } => r.groupId !== null,
-  )
-
-  // Helper map to de-duplicate identical segments
-  const segMap = new Map<string, { start: Point; end: Point }>()
-
-  for (let i = 0; i < groupedRects.length; i++) {
-    const a = groupedRects[i]
-    for (let j = i + 1; j < groupedRects.length; j++) {
-      const b = groupedRects[j]
-      if (a.groupId === b.groupId) continue // same merged rect → skip
-
-      // ── vertical shared edge (left/right) ─────────────────────────────
-      const aRight = a.x + a.width
-      const bRight = b.x + b.width
-      if (Math.abs(aRight - b.x) < TOL || Math.abs(bRight - a.x) < TOL) {
-        const x = Math.abs(aRight - b.x) < TOL ? aRight : bRight
-        const y0 = Math.max(a.y, b.y)
-        const y1 = Math.min(a.y + a.height, b.y + b.height)
-        if (y1 - y0 > TOL) {
-          const s = { x, y: y0 }
-          const e = { x, y: y1 }
-          segMap.set(getSegmentKey(s, e), { start: s, end: e })
-        }
-      }
-
-      // ── horizontal shared edge (top/bottom) ──────────────────────────
-      const aBot = a.y + a.height
-      const bBot = b.y + b.height
-      if (Math.abs(aBot - b.y) < TOL || Math.abs(bBot - a.y) < TOL) {
-        const y = Math.abs(aBot - b.y) < TOL ? aBot : bBot
-        const x0 = Math.max(a.x, b.x)
-        const x1 = Math.min(a.x + a.width, b.x + b.width)
-        if (x1 - x0 > TOL) {
-          const s = { x: x0, y }
-          const e = { x: x1, y }
-          segMap.set(getSegmentKey(s, e), { start: s, end: e })
-        }
-      }
-    }
-  }
-
-  // Convert map → outlineLines array
-  segMap.forEach(({ start, end }) => {
-    outlineLines.push({
-      id: `outline-${outlineLineId++}`,
-      start,
-      end,
-    })
-  })
-
-  const offsetPoint = (p: Point): Point => ({
-    x: p.x + offsetX,
-    y: p.y + offsetY,
-  })
-
-  const offsetLine = (l: Line): Line => ({
-    ...l,
-    start: offsetPoint(l.start),
-    end: offsetPoint(l.end),
-  })
-
-  const offsetRect = (r: CellContent): CellContent => ({
-    ...r,
-    x: r.x + offsetX,
-    y: r.y + offsetY,
-  })
-
-  return {
-    midlines: midlines.map(offsetLine),
-    allSegments: allSegments.map(offsetLine),
-    validSegments: validSegments.map(offsetLine),
-    mergedRectGroups: mergedRectGroups.map((g) => g.map(offsetRect)),
-    cellRects: cellContents.map(offsetRect),
-    gridRects: gridRects.map(offsetRect),
-    outlineLines: outlineLines.map(offsetLine),
-    // polygons: [], // Removed
-    // mergedPolygons: [], // Removed
-  }
 }
 
 const RECT_STAGE_ID = "rects" // new step-id for pre-merge rects
