+++ site/main2.tsx
@@ -1,17 +1,17 @@
-import React from "react";
-import { createRoot } from "react-dom/client";
-import CellBoundariesVisualization from "./claude-cell-boundaries";
+import React from "react"
+import { createRoot } from "react-dom/client"
+import CellBoundariesVisualization from "./cell-boundaries"
 
 const App: React.FC = () => {
-  return <CellBoundariesVisualization />;
-};
+  return <CellBoundariesVisualization />
+}
 
-const container = document.getElementById("root");
+const container = document.getElementById("root")
 if (container) {
-  const root = createRoot(container);
+  const root = createRoot(container)
   root.render(
     <React.StrictMode>
       <App />
     </React.StrictMode>,
-  );
+  )
 }
